_oteldemo/src/recommendationservice/requirements.txt
grpcio-health-checking==1.43.0
grpcio==1.51.3
opentelemetry-distro==0.36b0
opentelemetry-exporter-otlp-proto-grpc==1.15.0
python-dotenv==0.21.0
python-json-logger==2.0.4
psutil==5.9.2 # Importing this will also import opentelemetry-instrumentation-system-metrics when running opentelemetry-bootstrap
